# documents/edificad-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edificad API",
    "description": "Espeicificação da API Edificad, voltada para a gestão da ONG Edificando Vidas. ",
    "version": "0.0.1",
    "contact": {
      "name": "Repositório",
      "url": "https://github.com/jarnunes/edificad"
    },
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "https://edificad-production.up.railway.app/api",
      "description": "Produção"
    },
    {
      "url": "http://edificad.com:5000/api",
      "description": "Desenvolvimento"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticar um usuário e gerar token de acesso aos endpoints",
        "description": "Gera um token que será utilizado para acesso a todos os endpoints da API",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Autenticacao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacaoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Retornar array com usuários cadastrados",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioArray"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenInvalido"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastrar usuário",
        "description": "Endpoint para criação de um novo usuário. Roles aceitas ('WEBSERVICES', 'ADMIN', 'OPERATOR')",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/auth/{id}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Remover usuário referente ao ID do parâmetro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário a ser removido",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "example": "Usuário {username} removido com sucesso"
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "content": {
              "application/json": {
                "example": "Objeto não encontrado"
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Atualizar a senha",
        "security": [
          {
            "auth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Autenticacao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "example": "Senha atualizada com sucesso"
              }
            }
          }
        }
      }
    },
    "/auth/{email}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Obter usuário pelo email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      }
    },
...
    "/voluntario": {
      "get": {
        "tags": [
          "Voluntário"
        ],
        "summary": "Retorna um array com os voluntários cadastrados",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpf",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoluntarioArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Voluntário"
        ],
        "summary": "Cadastro de novo voluntário",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoluntarioPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": "Ocorreu um erro na validação do campo 'nome_campo'"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Voluntário"
        ],
        "summary": "Atualizar entidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Voluntario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          }
        }
      }
    },
    "/voluntario/{id}": {
      "get": {
        "tags": [
          "Voluntário"
        ],
        "summary": "Busca um voluntário pelo ID",
        "operationId": "getVoluntarioById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do voluntário ser retornada",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "content": {
              "application/json": {
                "example": "Objeto não encontrado"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Voluntário"
        ],
        "summary": "Remove o voluntário referente ao ID do parâmetro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do voluntário a ser removido",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "example": "Voluntário {nome_pessoa} removido com sucesso"
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "content": {
              "application/json": {
                "example": "Objeto não encontrado"
              }
            }
          }
        }
      }
    },
    "/voluntario/create-from-list": {
      "post": {
        "tags": [
          "Voluntário"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoluntarioPostArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoluntarioArray"
                }
              }
            }
          }
        }
      }
    },
    "/dependente": {
      "get": {
        "tags": [
          "Dependente"
        ],
        "summary": "Retorna um array com todos os dependentes cadastrados",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependenteArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dependente"
        ],
        "summary": "Criar um novo dependente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependentePost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dependente"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dependente"
        ],
        "summary": "Atualizar entidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dependente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "auth": []
    }
  ],
  "components": {
    "schemas": {
      "Autenticacao": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "marcelo129"
          },
          "password": {
            "type": "string",
            "example": "921elorem"
          }
        }
      },
      "AutenticacaoResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJQZXJzb24iLCJzdWIiOiJhZG1pbiIsImlkIjowLCJleHAiOjE2OTYxOTgwODN9.HLuHYcliYR2W33RgwMqkZPlxrIX1c_H6rmSmlLcruT8"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "full_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean",
            "example": false
          },
          "user_roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleUsuario"
            }
          }
        }
      },
      "UsuarioPost": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean",
            "example": false
          },
          "user_roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleUsuarioPost"
            }
          }
        }
      },
      "UsuarioArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Usuario"
        }
      },
      "RoleUsuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 22
          },
          "role": {
            "type": "string",
            "example": "ADMIN"
          }
        }
      },
...
      "RoleUsuarioPost": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "WEBSERVICES"
          }
        }
      },
...
      "Voluntario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Marcello Du Claro"
          },
          "telefone": {
            "type": "string",
            "minLength": 11,
            "maxLength": 15,
            "example": "(99) 99999-9999"
          },
          "email": {
            "type": "string",
            "example": "marcelo.duclaro@edificad.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "endereco": {
            "$ref": "#/components/schemas/Endereco"
          }
        }
      },
      "VoluntarioArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Voluntario"
        }
      },
      "VoluntarioPost": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Marcello Du Claro"
          },
          "telefone": {
            "type": "string",
            "minLength": 11,
            "maxLength": 15,
            "example": "(99) 99999-9999"
          },
          "email": {
            "type": "string",
            "example": "marcelo.duclaro@edificad.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "endereco": {
            "$ref": "#/components/schemas/EnderecoPost"
          }
        }
      },
      "VoluntarioPostArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/VoluntarioPost"
        }
      },
...
      "Dependente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "data_nascimento": {
            "type": "string"
          },
          "id_responsavel": {
            "type": "integer"
          }
        }
      },
      "DependenteArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Dependente"
        }
      },
      "DependentePost": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "data_nascimento": {
            "type": "string"
          },
          "id_responsavel": {
            "type": "integer"
          }
        }
      },
...
      "Endereco": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "logradouro": {
            "type": "string",
            "example": "Rua Alvarenga Dia"
          },
          "numero": {
            "type": "string",
            "example": "952 Casa C"
          },
          "cep": {
            "type": "string",
            "example": "00000-678"
          },
          "bairro": {
            "type": "string",
            "example": "Bairro Alv"
          },
          "cidade": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "estado": {
            "type": "string",
            "example": "MG"
          }
        }
      },
      "EnderecoPost": {
        "type": "object",
        "properties": {
          "logradouro": {
            "type": "string",
            "example": "Rua Alvarenga Dia"
          },
          "numero": {
            "type": "string",
            "example": "952 Casa C"
          },
          "cep": {
            "type": "string",
            "example": "00000-678"
          },
          "bairro": {
            "type": "string",
            "example": "Bairro Alv"
          },
          "cidade": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "estado": {
            "type": "string",
            "example": "MG"
          }
        }
      },
...
      "TokenInvalido": {
        "type": "object",
        "properties": {
          "cause": {
            "type": "string",
            "example": "Invalid token"
          },
          "message_error": {
            "type": "string",
            "example": "The Token has expired on 2023-10-16T18:43:02Z."
          },
          "path": {
            "type": "string",
            "example": "/api/cesta"
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}